Refactor retrieve data to place it in data.py module
